+++ wp-content/themes/intranet/views/partials/header/intranet.blade.php
@@ -4,7 +4,7 @@
             <div class="grid-xs-12 {!! apply_filters('Municipio/header_grid_size','grid-md-12'); !!}">
                 <div class="grid">
                     <div class="grid-xs-12">
-                        <a href="{{ network_home_url() }}" class="site-title"><h1 class="no-margin no-padding">Helsingborg stads intranät</h1></a>
+                        <a href="{{ network_home_url() }}" class="site-title"><h1 class="no-margin no-padding">{{ get_site_option('site_name') }}</h1></a>
 
                         <form class="search" method="get" action="{{ home_url() }}">
                             <label for="searchkeyword-0" class="sr-only">{{ get_field('search_label_text', 'option') ? get_field('search_label_text', 'option') : __('Search', 'municipio') }}</label>