+++ changelog.md
@@ -7,7 +7,8 @@ Version 0.2.3 (tbd)
 ===================
 - Require Active Directory Integration (AD1) plugin
 - Adds "edit" and "remove" actions to the user systems admin
-- 
+- Added shortcode "explain" to Municipio-theme for inserting questionmarks with tooltips in post_content
+- Adds metabox to pages for setting custom table of contents title
 
 Version 0.2.2 (2016-06-29)
 ==========================